REF: more shift and diff tests
